Revert back to using proj_points for colorgrade
It was a bug in the submodule that I worked hours
to identify and patch. Then when I was about to
open a PR, I noticed it was already patched a few
days ago... classic:
   https://github.com/cwendt94/espn-api/commit/87c21ffe4e1b0f0c1c4754f4c738a7c7486d981a

Anyway, projected points is once again accurate and
does not randomly result in a different number.

Signed-off-by: xNickChildx <nnac123@gmail.com>